requirements.txt
aenum==3.1.15
can-isotp==1.8
pyserial==3.5
python-can==3.3.4
udsoncan==1.17
windows-curses==2.3.2
wrapt==1.16.0
